Updated data stored in Core Data
Removed the basic charge attribute. It could be stored and could be changed in settings. But it isn't needed anymore because it is an irrelevant setting.

Added a new settings which stores if the splash screen was already finished. If so the app can be normally used. If not you are shown some splash screens which require the user to state some settings.