Big performance improvments for graph
Big performance improvments for the graph on the main tab. This was especially a problem on not current devices like iPhone 12 lineup and iPhone 11 lineup.

Technical:
All views of the graph are now combined into a single view and than rendered. This helps to boost performance. Also the Metal API is now used directly.