Now stores values in Cheapest Time
Now stores values entered in the "Power" and the "Total consumption" field to persistent storage. The next time the user enteres some data in those fields, closes the app completely and reopens it, the app will read the stored values and pre-enter them. This allows a more faster usage of the app for the user.