UIDatePicker replaces SwiftUI DatePicker
Due to multiple issues with the SwiftUI DatePicker (which is based on UIDatePicker) the app now directly uses UIDatePicker's. Those appear (until now) only on the Cheapest Time page. This migration now allows to set multiple UIDatePicker settings which couldn't all be set through SwiftUI's implementation of DatePicker. This solves multiple issues which occurred in the app.