Store APNs token locally on device
The last apns token which could be sent to the server and is valid is now stored on the device.
APNs tokens for a device with AWattPrice can change. Therefor each time the app is launched it has to check if the current APNs token (provided by Apple) is still the currently stored token on the backend. If not the token is sent again to the backend.