Added scheduled updating script
Scheduled updates are invoked by a service like cron and run this newly added script.
When this python script is run it calls the AWattPrice backend.
This will start a request process on the backend and will check if there are new prices available to download.

This is needed to have current cached data and to send push notifications if new data is available (push notifications not yet implemented but will be handled by backend (fastapi part)).