Updating bug fixes
Fixed bug where it could happen that the aWATTar API is polled twice or even more in a time range - more often as speicified in settings as the minimum time between updates.
This would have happened:
-> Multiple clients connect at the same time -> all tasks read current data on server -> all tasks see that they need to update -> all update.

This would happen now:
-> Multiple clients connect at the same time -> one client gets access to check if they need to update -> needs to update -> updates | other clients m
ust wait in this time -> update finishes and is written to local drive -> other process are allowed to continue -> read fresh data and don't update ag
ain.