notifications: s/==/is/
True, None, False are objects instantiated by the Python
interpreter as singletons. IOW, there is one True object
at runtime.

Example:

In [1]: id(True)
Out[1]: 4364119696

In [2]: id(1 == 1)
Out[2]: 4364119696

That is why one should use the keyword "is" to compare if
a boolean expression yields e. g. True. Comparing e. g. numbers
is done via ==. Strictly speaking 1000 and 1000 are equal but
not the same.