backend poll: Mixed up types hid a bug
uuid.uuid4() generates a random uuid4. The name has
it already that it is a unique ID.

What happened here was that uuid.uuid4() generates
a UUID object. However, data.meta.uuid is a str.
Hence, the line new_uuid = uuid.uuid4().hex was
doing the actual assignment of a new uuid.

I'd say it is safe to assume that uuid4 generates a
unique ID.