For Widget: Added time zone modifi function.
Added a function which gets the current time of a timezone and allows to modify the hour, minute or seconds of this time. Returned is a Date object which reflects this modified date. Example: Get Europe/Berlin time, set hour to 12, minute to 30 and second to 0, then convert back to Date object which reflects the new/modified Europe/Berlin time as UTC time.