Update get energy data
Update function to get the current energy data. This function checks if it can get the current data from cache. If so, it checks if this data is still valid. Only if no data exists or data is outdated, it polls from the backend. In common scenarios this can reduce data usage to at least -(80-90%).