Fixed issue where apps database wouldn't be copied to the app group.
This was due to the widget extension being executed before the app, even if no widget is added to the home screen. iOS just tries to preload the widget configuration etc. to quickly display it in the widget gallery. Because of this the widget extension created a database inside of the app group. As soon as the app launched it would detect the database inside the app group and wouldn't attempt to copy its database to the app group anymore.